refactor: Member 패키지 이동 (#15)
