+++ src/main/java/com/octagram/pollet/member/infrastructure/MemberRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.octagram.pollet.member.infrastructure;
 
-import com.octagram.pollet.member.domain.Member;
+import com.octagram.pollet.member.domain.model.Member;
 
 public interface MemberRepositoryCustom {
